requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
customtkinter
ddgs==9.13.1
duckduckgo_search==8.1.1
google-genai
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.4
ollama==0.6.1
openai
primp==1.2.2
pydantic==2.13.0
pydantic_core==2.46.0
typing-inspection==0.4.2
typing_extensions==4.15.0
winotify==1.1.0
